backend/pyproject.toml
[project]
name = "vectorforge-backend"
version = "0.1.0"
description = "VectorForge - Raster to Vector conversion micro-SaaS"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13",
    "vtracer>=0.6",
    "svgpathtools>=1.6",
    "ezdxf>=0.19",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    "pillow>=10.0",
    "svglib>=0.19",
    "reportlab>=4.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
